Fix local codelist encoding issue
The csv files in local_codelists had been created with the
UTF-8 with BOM (byte order marker) encoding.

codelist_from_csv() expects UTF-8 without BOM encoding.

convert_utf8_bom.py reads every csv file in local_codelists,
checks to see if it is in utf8-with-bom encoding, and if so,
converts it to utf-without-bom format.

If any new codelists are added to local_codelists that were
created by the same process as the previous problematic codelists,
this script will need to be run again.